Format Playlist popups + icons, format top bar on user page
Formats the playlist popups (and other popups probably), plus playlist
icons (may have made a change that will break other arrow dropdowns, if
this is the case, will need to go in and explicitly turn off the filter
for the playlist dropdowns).  Also formatted the secondary bar that
shows on your own user page.